refactor(providers): Codex cloud as a plugin
The one CLI-observed provider, and the one provider write through a CLI.
`codexCloudPlugin` is a value over `cliPass`: `observe`/`latest` from the
pass, `connection()` for the settings row, `projects()` from the
environment sweep, and `acts: { createWorkspace }` and nothing else —
Codex documents no way to message or steer a task already running, so
its cloud sessions name no session act and the honest absence stands.

`cloud-adapter.ts` splits into `cloud-wire.ts` (the fixed invocation
table, the task shapes, the state map) and `cloud-observe.ts` (the sweep
and the observation), leaving `cloud.ts` as the plugin.

Two more base classes go: `CliSessionAdapter`, whose only subclass this
was, and `CompositeSessionProviderAdapter`, replaced in `registrations`
by `mergePlugins`, with the "a pass fails whole" reasoning carried over
in the function that now holds it. `CODEX_ADAPTER_DEFAULTS` stood for
two different tables in two files; the cloud one is
`CODEX_CLOUD_DEFAULTS` now.

The contract suite's creation case asks through `dispatchAct` rather
than reaching a handler directly, because for a plugin that is the only
route an act has: the resolution that refuses an unreported project is
`dispatchAct`'s, and asking the handler behind it tested a door nothing
in production opens.
