+++ packages/session/src/index.ts
@@ -3,7 +3,6 @@ export * from "./act-results.js";
 export * from "./advertised-acts.js";
 export * from "./agent-identities.js";
 export * from "./bounds.js";
-export * from "./composite-provider-adapter.js";
 export * from "./normalize.js";
 export * from "./provider-contract.js";
 export * from "./provider-identity.js";
